+++ app/src/main/java/com/example/eventmatics/Events_Data_Holder_Activity/VendorDataHolderActivity.kt
@@ -25,6 +25,7 @@ import com.example.eventmatics.SQLiteDatabase.Dataclass.DatabaseAdapter.LocalDat
 import com.example.eventmatics.SQLiteDatabase.Dataclass.Vendor
 import com.example.eventmatics.SwipeGesture.BudgetSwipeToDelete
 import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.android.material.snackbar.BaseTransientBottomBar
 import com.google.android.material.snackbar.Snackbar
@@ -119,31 +120,20 @@ class VendorDataHolderActivity : AppCompatActivity(),VendorDataHolderClass.onIte
                 ItemTouchHelper.LEFT->{
                 if(position!=RecyclerView.NO_POSITION){
                     val deleteditem=vendorlist[position]
-                    db.deleteVendor(deleteditem)
-                    adapter.notifyItemRemoved(position)
-                    val snackbar=Snackbar.make(this@VendorDataHolderActivity.recyclerView,"Vendor Data Deleted",Snackbar.LENGTH_LONG)
-                        .addCallback(object:BaseTransientBottomBar.BaseCallback<Snackbar>(){
-                            override fun onDismissed(transientBottomBar: Snackbar?, event: Int) {
-                                super.onDismissed(transientBottomBar, event)
-                                recreate()
-                            }
 
-                            override fun onShown(transientBottomBar: Snackbar?) {
-                                super.onShown(transientBottomBar)
-                                transientBottomBar?.setAction("UNDO"){
-                                    vendorlist.add(position,deleteditem)
-                                    adapter.notifyItemInserted(position)
-                                }
-                            }
-                        }).apply {
-                            animationMode=Snackbar.ANIMATION_MODE_FADE
+                    adapter.notifyItemRemoved(position)
+                    MaterialAlertDialogBuilder(this@VendorDataHolderActivity)
+                        .setTitle("Delete Item")
+                        .setMessage("Do you want to delete this item?")
+                        .setPositiveButton("Delete"){dialog,_->
+                            db.deleteVendor(deleteditem)
+                            recreate()
                         }
-                    snackbar.setActionTextColor(
-                        ContextCompat.getColor(
-                            this@VendorDataHolderActivity, androidx.browser.R.color.browser_actions_bg_grey
-                        )
-                    )
-                    snackbar.show()
+                        .setNegativeButton("Cancel"){dialog,_->
+                            dialog.dismiss()
+                            recreate()
+                        }
+                        .show()
                 }
             }
                     ItemTouchHelper.RIGHT->{
@@ -162,48 +152,40 @@ class VendorDataHolderActivity : AppCompatActivity(),VendorDataHolderClass.onIte
                                 newStatus="Paid"
                                 NewBalance=Balance.toFloat()
                             }
-                            val rowaffected=db.updateVendorPaid(vendor.id,newStatus,"$NewBalance")
-                            if(rowaffected>0){
-                                val snackbar=Snackbar.make(this@VendorDataHolderActivity.recyclerView,"Vendor Data Updated",Snackbar.LENGTH_LONG)
-                                    .addCallback(object:BaseTransientBottomBar.BaseCallback<Snackbar>(){
-                                        override fun onDismissed(
-                                            transientBottomBar: Snackbar?,
-                                            event: Int
-                                        ) {
-                                            super.onDismissed(transientBottomBar, event)
-                                            recreate()
-                                        }
-                                        override fun onShown(transientBottomBar: Snackbar?) {
-                                            transientBottomBar?.setAction("UNDO"){
-                                                if (newStatus=="Paid"){
-                                                    previousstatus="Not Paid"
-                                                    NewBalance=vendor.estimatedAmount.toFloat()
-                                                }
-                                                else{
-                                                    previousstatus="Paid"
-                                                    NewBalance=Balance.toFloat()
-                                                }
-                                                db.updateVendorPaid(vendor.id,previousstatus,"$NewBalance")
-                                                recreate()
-                                            }
+                            when(newStatus){
+                                "Paid"->MaterialAlertDialogBuilder(this@VendorDataHolderActivity)
+                                    .setTitle("Vendor Payment Status")
+                                    .setMessage("Is Vendor Payment Done")
+                                    .setPositiveButton("Yes"){dialog,_->
+                                        db.updateVendorPaid(vendor.id,newStatus,"$NewBalance")
+                                        recreate()
 
-                                            super.onShown(transientBottomBar)
-                                        }
-                                    }).apply {
-                                        animationMode=Snackbar.ANIMATION_MODE_FADE
                                     }
-                                snackbar.setActionTextColor(
-                                    ContextCompat.getColor(
-                                        this@VendorDataHolderActivity, androidx.browser.R.color.browser_actions_bg_grey
-                                    )
-                                )
-                                snackbar.show()
-                            }
-                        }
-                    }
+                                    .setNeutralButton("No"){dialog,_->
+                                        dialog.dismiss()
+                                        recreate()
+                                    }
+                                    .show()
+                                "Not Paid"->MaterialAlertDialogBuilder(this@VendorDataHolderActivity)
+                                    .setTitle("Vendor Payment Status")
+                                    .setMessage("Is Vendor Payment Not Done")
+                                    .setPositiveButton("Yes"){dialog,_->
+                                        db.updateVendorPaid(vendor.id,newStatus,"$NewBalance")
+                                        recreate()
 
-        }}}
+                                    }
+                                    .setNeutralButton("No"){dialog,_->
+                                        dialog.dismiss()
+                                        recreate()
+                                    }
+                                    .show()
+                            }
+
+                        }
+                    } }}
+        }
         val itemtouch=ItemTouchHelper(swipe)
+
         itemtouch.attachToRecyclerView(recyclerView)
     }
